refactor: split manifest, consolidate validation, race-test clean
- Extract Manifest() and Actions() to manifest.go (amadeus.go: 473→312 lines)
- Consolidate validCabins, validGenders, and all validation constants
  into validate.go (single source of truth for input validation)
- Verified no race conditions with `go test -race`
- All 49 tests pass
