+++ connectors/amadeus/validate.go
@@ -19,16 +19,32 @@ func validIATACode(s string) bool {
 	return iataCodeRe.MatchString(s)
 }
 
-// maxFlightOfferBytes is the maximum size of a flight_offer JSON blob
-// we'll accept. Prevents forwarding arbitrarily large payloads to Amadeus.
-const maxFlightOfferBytes = 100 * 1024 // 100 KB
+// validCabins is the set of allowed cabin classes for flight search.
+var validCabins = map[string]bool{
+	"ECONOMY":         true,
+	"PREMIUM_ECONOMY": true,
+	"BUSINESS":        true,
+	"FIRST":           true,
+}
 
-// maxTravelers is the maximum number of travelers allowed in a booking,
-// matching the Amadeus API limit for adults in a single PNR.
-const maxTravelers = 9
+// validGenders is the set of allowed gender values for Amadeus traveler records.
+var validGenders = map[string]bool{
+	"MALE":   true,
+	"FEMALE": true,
+}
 
-// maxRemarkLen is the maximum length of a booking remark.
-const maxRemarkLen = 500
+const (
+	// maxFlightOfferBytes is the maximum size of a flight_offer JSON blob
+	// we'll accept. Prevents forwarding arbitrarily large payloads to Amadeus.
+	maxFlightOfferBytes = 100 * 1024 // 100 KB
 
-// maxResultsCap is the upper bound for max_results in search actions.
-const maxResultsCap = 250
+	// maxTravelers is the maximum number of travelers allowed in a booking,
+	// matching the Amadeus API limit for adults in a single PNR.
+	maxTravelers = 9
+
+	// maxRemarkLen is the maximum length of a booking remark.
+	maxRemarkLen = 500
+
+	// maxResultsCap is the upper bound for max_results in search actions.
+	maxResultsCap = 250
+)
